+++ tests/PSRule.Tests/AssertTests.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using PSRule.Pipeline;
-using System;
 using System.Management.Automation;
 using Xunit;